fix: smarter routing, less duplicates
