+++ frontend/src/app/[locale]/[mode]/(mypages-sections)/arenden/[caseId]/[[...currentTab]]/page.tsx
@@ -1,13 +1,4 @@
 import CaseTabLayout, { CaseCurrentTab } from '@layouts/pages/mypages-sections/cases/case/case-tab-layout.component';
-import { appName } from '@utils/app-name';
-import { capitalize } from 'lodash';
-
-export async function generateMetadata({ params }: { params: Promise<{ caseId: string; currentTab: string }> }) {
-  const { currentTab } = await params;
-  return {
-    title: `${capitalize(currentTab || 'uppgifter')} - Ärende - Företag - ${appName()}`,
-  };
-}
 
 export default async function CurrentTab({ params }) {
   const { caseId, currentTab } = await params;
