+++ frontend/src/app/[locale]/[mode]/(mypages-sections)/arenden/[caseId]/layout.tsx
