+++ frontend/src/app/[locale]/[mode]/(mypages-sections)/beslut-och-dokument/[assetId]/layout.tsx
