+++ frontend/src/app/[locale]/[mode]/(mypages-sections)/beslut-och-dokument/[assetId]/page.tsx
@@ -1,0 +1,5 @@
+import Asset from '@layouts/pages/mypages-sections/decicions-and-documents/asset.component';
+
+export default function Index() {
+  return <Asset />;
+}
