+++ frontend/src/app/[locale]/[mode]/(mypages-sections)/beslut-och-dokument/page.tsx
@@ -1,12 +1,5 @@
 import { PagesLayout } from '@layouts/pages-layout.component';
 import { DecisionsAndDocuments } from '@layouts/pages/mypages-sections/decicions-and-documents/decisions-and-documents.component';
-import { appName } from '@utils/app-name';
-
-export async function generateMetadata() {
-  return {
-    title: `Beslut och dokument - Privat - ${appName()}`,
-  };
-}
 
 export default function Index() {
   return (
