+++ frontend/src/app/[locale]/[mode]/(mypages-sections)/fakturor/page.tsx
@@ -1,12 +1,5 @@
 import { PagesLayout } from '@layouts/pages-layout.component';
 import Invoices from '@layouts/pages/mypages-sections/invoices.component';
-import { appName } from '@utils/app-name';
-
-export async function generateMetadata() {
-  return {
-    title: `Fakturor - Privat - ${appName()}`,
-  };
-}
 
 export default function Fakturor() {
   return (
