+++ frontend/src/app/[locale]/[mode]/(mypages-sections)/layout.tsx
@@ -6,9 +6,11 @@ import { useAppContext } from '../../../../contexts/app.context';
 import { RepresentingEntity, RepresentingMode } from '../../../../interfaces/app';
 import { DefaultLayout } from '../../../../layouts/default-layout.component';
 import { useApi } from '../../../../services/api-service';
+import { useParams } from 'next/navigation';
 
 export default function Layout({ children }) {
-  const { setRepresentingMode, representingMode } = useAppContext();
+  const { representingMode, setRepresentingMode } = useAppContext();
+  const { mode } = useParams();
   const {
     data: representingEntity,
     isLoading: representingIsLoading,
@@ -25,14 +27,27 @@ export default function Layout({ children }) {
 
   useEffect(() => {
     if (mounted && !representingIsLoading && !representingIsFetching) {
-      if (representingEntity?.mode !== RepresentingMode.BUSINESS || representingMode !== RepresentingMode.BUSINESS) {
+      if (
+        mode === 'foretag' &&
+        (representingEntity?.mode !== RepresentingMode.BUSINESS || representingMode !== RepresentingMode.BUSINESS)
+      ) {
         setRepresentingMode(RepresentingMode.BUSINESS);
+      } else if (
+        mode === 'privat' &&
+        (representingEntity?.mode !== RepresentingMode.PRIVATE || representingMode !== RepresentingMode.PRIVATE)
+      ) {
+        setRepresentingMode(RepresentingMode.PRIVATE);
       }
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [mounted, representingIsLoading, representingIsFetching]);
 
-  if (!mounted || representingEntity === undefined || representingEntity.BUSINESS === undefined) {
+  if (
+    !mounted ||
+    representingEntity === undefined ||
+    (mode === 'foretag' && representingEntity.BUSINESS === undefined) ||
+    (mode === 'privat' && representingEntity.PRIVATE === undefined)
+  ) {
     return <FullscreenMainSpinner />;
   }
