+++ frontend/src/app/[locale]/[mode]/(mypages-sections)/oversikt/page.tsx
@@ -1,12 +1,5 @@
 import { PagesLayout } from '@layouts/pages-layout.component';
 import Overview from '@layouts/pages/mypages-sections/overview.component';
-import { appName } from '@utils/app-name';
-
-export async function generateMetadata() {
-  return {
-    title: `Översikt - Privat - ${appName()}`,
-  };
-}
 
 export default function Oversikt() {
   return (
