+++ frontend/src/app/[locale]/[mode]/(mypages-sections)/profil/page.tsx
@@ -1,12 +1,5 @@
 import { PagesLayout } from '@layouts/pages-layout.component';
 import { Profile } from '@layouts/pages/mypages-sections/profile/profile.component';
-import { appName } from '@utils/app-name';
-
-export async function generateMetadata() {
-  return {
-    title: `Profil - Privat - ${appName()}`,
-  };
-}
 
 export default function Profil() {
   return (
