+++ frontend/src/app/[locale]/[mode]/valj-foretag/page.tsx
@@ -1,0 +1,5 @@
+import ValjForetag from '@layouts/pages/valj-foretag/valj-foretag.component';
+
+export default function Index() {
+  return <ValjForetag />;
+}
