--- frontend/src/app/[locale]/foretag/(mypages-sections)/arenden/page.tsx
@@ -1,17 +1,0 @@
-import { PagesLayout } from '@layouts/pages-layout.component';
-import Cases from '@layouts/pages/mypages-sections/cases.component';
-import { appName } from '@utils/app-name';
-
-export async function generateMetadata() {
-  return {
-    title: `Ärenden - Företag - ${appName()}`,
-  };
-}
-
-export default function Arenden() {
-  return (
-    <PagesLayout>
-      <Cases />
-    </PagesLayout>
-  );
-}
