--- frontend/src/app/[locale]/foretag/(mypages-sections)/beslut-och-dokument/[assetId]/layout.tsx
@@ -1,6 +1,0 @@
-import AssetLayout from '@layouts/pages/mypages-sections/decicions-and-documents/asset-layout.component';
-
-export default async function layout({ params, children }) {
-  const { assetId } = await params;
-  return <AssetLayout assetId={assetId}>{children}</AssetLayout>;
-}
