--- frontend/src/app/[locale]/foretag/(mypages-sections)/fakturor/page.tsx
@@ -1,17 +1,0 @@
-import { PagesLayout } from '@layouts/pages-layout.component';
-import Invoices from '@layouts/pages/mypages-sections/invoices.component';
-import { appName } from '@utils/app-name';
-
-export async function generateMetadata() {
-  return {
-    title: `Fakturor - Företag - ${appName()}`,
-  };
-}
-
-export default function Fakturor() {
-  return (
-    <PagesLayout>
-      <Invoices />
-    </PagesLayout>
-  );
-}
