--- frontend/src/app/[locale]/foretag/valj-foretag/page.tsx
@@ -1,12 +1,0 @@
-import ValjForetag from '@layouts/pages/valj-foretag/valj-foretag.component';
-import { appName } from '@utils/app-name';
-
-export async function generateMetadata() {
-  return {
-    title: `Välj företag - Företag - ${appName()}`,
-  };
-}
-
-export default function Index() {
-  return <ValjForetag />;
-}
