--- frontend/src/app/[locale]/privat/beslut-och-dokument/[assetId]/page.tsx
@@ -1,12 +1,0 @@
-import Asset from '@layouts/pages/mypages-sections/decicions-and-documents/asset.component';
-import { appName } from '@utils/app-name';
-
-export async function generateMetadata() {
-  return {
-    title: `Beslut och dokument - Privat - ${appName()}`,
-  };
-}
-
-export default function Index() {
-  return <Asset />;
-}
