--- frontend/src/app/[locale]/privat/page.tsx
@@ -1,7 +1,0 @@
-import { RepresentingMode } from '@interfaces/app';
-import { redirect } from 'next/navigation';
-import { getRepresentingModeRoute } from '../../../utils/representingModeRoute';
-
-export default function Index() {
-  redirect(`${getRepresentingModeRoute(RepresentingMode.PRIVATE)}/oversikt`);
-}
